Added reactstrap to style
