+++ team-builder/src/style/style.css
@@ -1,0 +1,4 @@
+.container{
+    
+    border:3px solid white;
+}
